Move Parser into bob.parser pkg
